feat: Refactor API response

+++ src/services/columnService.js
@@ -1,65 +1,50 @@
 import { columnModel } from '~/models/columnModel'
 import { boardModel } from '~/models/boardModel'
 import { cardModel } from '~/models/cardModel'
-import ApiError from '~/utils/ApiError'
-import { StatusCodes } from 'http-status-codes'
+import { GET_CLIENT } from '~/config/mongodb'
 
-const createNew = async reqBody => {
-  try {
-    const newColumn = {
-      ...reqBody
-    }
-    const createColumn = await columnModel.createNew(newColumn)
-    const getNewColumn = await columnModel.findOneById(createColumn.insertedId)
+const createNew = async dataInput => {
+  const createColumn = await columnModel.createNew({
+    ...dataInput
+  })
 
-    if (getNewColumn) {
-      getNewColumn.cards = []
-    }
+  const getNewColumn = await columnModel.findOneById(createColumn.insertedId)
 
-    await boardModel.pushColumnOrderIds(getNewColumn)
-
-    return getNewColumn
-  } catch (error) {
-    throw new Error(error)
+  if (getNewColumn) {
+    getNewColumn.cards = []
   }
+
+  await boardModel.pushColumnOrderIds(getNewColumn)
+
+  return getNewColumn
 }
 
-const update = async (columnId, reqBody) => {
-  try {
-    const updateData = {
-      ...reqBody,
-      updatedAt: Date.now()
-    }
-    return await columnModel.update(columnId, updateData)
-  } catch (error) {
-    throw new Error(error)
-  }
+const update = async (columnId, dataInput) => {
+  return await columnModel.update(columnId, {
+    ...dataInput,
+    updatedAt: Date.now()
+  })
 }
 
 const deleteItem = async columnId => {
+  const targetColumn = await columnModel.findOneById(columnId)
+
+  if (!targetColumn || !targetColumn.boardId) {
+    return { deleteResult: 'Column already deleted or not found!' }
+  }
+
+  const session = GET_CLIENT().startSession()
   try {
-    const targetColumn = await columnModel.findOneById(columnId)
-
-    if (!targetColumn) {
-      return { deleteResult: 'Column already deleted or not found!' }
-    }
-
-    await cardModel.deleteMany(columnId)
-
-    await boardModel.deleteColumnOrderIds(targetColumn)
-
-    const deleteResult = await columnModel.deleteItem(columnId)
-
-    if (deleteResult.deletedCount === 0) {
-      throw new ApiError(StatusCodes.NOT_FOUND, 'Column not found or already deleted!')
-    }
-
-    return { deleteResult: 'Column and its Cards deleted successfully!' }
+    await session.withTransaction(async () => {
+      await cardModel.deleteMany(columnId)
+      await boardModel.deleteColumnOrderIds(targetColumn)
+      await columnModel.deleteItem(columnId)
+    })
   } catch (error) {
-    if (error instanceof ApiError) {
-      throw error
-    }
+    session.abortTransaction()
     throw new Error(error)
+  } finally {
+    await session.endSession()
   }
 }
